app/requirements/base.txt
argon2-cffi==21.3.0
django==4.1.1
django-anymail==8.6
django-filter==22.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-jazzmin==2.5.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
whitenoise==6.2.0
psycopg2-binary==2.9.3
drf_spectacular==0.24.2
